+++ lib/main.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'quote.dart';
+import 'quote_card.dart';
 
 void main() => runApp(MaterialApp(
   home: QuoteList()
@@ -15,14 +16,10 @@ class QuoteList extends StatefulWidget {
 
 class _QuoteListState extends State<QuoteList> {
   List<Quote> quotes = [
-    Quote(author: "Zenyatta", text: "Repetition is the path to mastery x"),
-    Quote(author: "Guo Songx", text: "You cannot after two horses at the same time a"),
-    Quote(author: "Zenyatta", text: "The pain is the best teacher d"),
+    Quote(author: "Zenyatta", text: "Repetition is the path to mastery"),
+    Quote(author: "Guo Songx", text: "You cannot after two horses at the same time"),
+    Quote(author: "Zenyatta", text: "The pain is the best teacher soo"),
   ];
-
-  Widget quoteTemplate(quote) {
-    return QuoteCard(quote: quote);
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -38,45 +35,7 @@ class _QuoteListState extends State<QuoteList> {
         /*children: quotes.map((quote){
           return Text(quote);
         }).toList()*/
-        children: quotes.map((quote)=>quoteTemplate(quote)).toList(),
-      ),
-    );
-  }
-}
-
-class QuoteCard extends StatelessWidget {
-
-  final Quote quote;
-  QuoteCard({this.quote});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card (
-      margin: EdgeInsets.fromLTRB(16.0, 16.0, 16.0, 0),
-      color: Colors.deepPurpleAccent,
-      child: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            Text(
-              quote.text,
-              style: TextStyle(
-                fontSize: 18.0,
-                color:Colors.white
-              )
-            ),
-            SizedBox(height: 6.0),
-            Text(
-              quote.author,
-              textAlign: TextAlign.right,
-              style: TextStyle(
-                fontSize: 14.0,
-                color: Colors.lightGreen,
-              ),
-            ),
-          ],
-        ),
+        children: quotes.map((quote)=> QuoteCard(quote: quote)).toList(),
       ),
     );
   }
